# Fixed reschedule issue in Kafka