@oandrew 2020-05-22
Thank you! I was just about to make a PR fixing this same issue. In our case 500ms reschedule was limiting rate to about 300k rows per second. With a similar fix - 1-2M+/s