@azat 2020-05-24
Also seems that now the upstream cppkafka can be used (your PR is merged, and other build fixes already there too)