@alexey-milovidov 2020-05-25
@filimonov Do you have any ideas about test_storage_kafka/test.py::test_kafka_flush_by_block_size test fail?