@filimonov 2020-05-25
Can't reproduce locally. Most probably the PR just pushed a strange balance of sleeps inside those Kafka tests, and some sleeps need to be reviewed/adjusted. Locally i have another test flapping now.