@filimonov 2020-05-25
> Also seems that now the upstream cppkafka can be used (your PR is merged, and other build fixes already there too)

Not in that PR. Cppkafka has some other changes too. That PR is a bugfix, want to keep it compact and limit side effects.