@filimonov 2020-05-25
> I would also add details about the performance tests (and notes from the PR description) into the commit itself
And also maybe it worth updating title of the PR so that it includes the possible "performance" benefits.

Changelog entry contains that. Commit messages - discussed on telegram (it's a nice practice, but has no sence if only one people will follow that, and anyway it can't replace PR description and documentation in PR/tasks etc)