@filimonov 2020-05-26
Hm. It looks like a real issue (minor).

It can exit the poll loop prematurely by eof, w/o retries. One line fix.

Will add one more test for that and fix it here too.