@filimonov 2020-05-27
Rebased on the newer master (was on the top of broken one), squashed.

Test fixed, another bug discovered https://github.com/ClickHouse/ClickHouse/pull/11149
(Will fix the bug in follow-up PR)