@filimonov 2020-05-27
Did a performance tests:

Test https://github.com/filimonov/ch-kafka-consume-perftest

Format TSV
Topic with 4 partitions, prefilled with 33554432 records.

Consuming time with kafka_num_consumers=2 measured.

| kafka_max_block_size | consume_time  (s, before) | consume_speed (rec/sec, before) | consume_time (s, that PR) | consume_speed (rec/sec, that PR) | improvement |
|-:|-:|-:|-:|-:|-:|
|                65536 |          174 |        192842 |           99 |        338934 | +75.76% | 
|               131072 |          130 |        258111 |           97 |        345922 | +34.02% | 
|               262144 |           88 |        381300 |           64 |        524288 | +37.50% | 
|               524288 |           66 |        508400 |           68 |        493448 | -2.94%  | 
|              1048576 |           65 |        516222 |           60 |        559241 | +8.33%  |